[PAY-549] Write out IP with each call to relay
